Limit the offline image-kind error and isolate the hive native API
